requirements/requirements.txt
matplotlib
numpy
scipy
pillow
scikit-image
scikit-learn
h5py
pandas
opencv-python
schema
tqdm
pyyaml
tabulate
ipykernel
opencv-python
wandb
bm3d
SciencePlots
cvxopt
ax-platform

tensorflow-gpu<2.11
tensorflow-addons
tensorflow-gan
tensorflow-datasets
tensorflow-estimator